created component for drawer and Map so the dynamics can happen there
